requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and data handling
sqlite-utils==3.35.2
pydantic==2.5.0

# AI and embeddings
openai==1.3.7
tiktoken==0.5.2
numpy==1.24.3

# Voice and audio
elevenlabs==0.2.26
pyaudio==0.2.11
speechrecognition==3.10.0

# Integrations
slack-sdk==3.23.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
notion-client==2.2.1

# Utilities
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
asyncio-mqtt==0.16.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1